feat: Visualize Line of Contact (LBS) on map
Adds a visual representation of the Line of Contact (LBS) to the interactive region map. This change updates the `api/get-conflict-zones.ts` endpoint to return GeoJSON data including a LineString for the LBS and modifies the `components/InteractiveRegionMap.tsx` component to render this LineString with a distinct red style. The map legend has also been updated to include an entry for the LBS.